Move wiki editor into cube
