+++ cube/package.json
@@ -18,10 +18,6 @@
       "types": "./src/react/index.tsx",
       "default": "./src/react/index.tsx"
     },
-    "./editor": {
-      "types": "./src/editor/index.ts",
-      "default": "./src/editor/index.ts"
-    },
     "./mcp": {
       "types": "./src/mcp/index.ts",
       "default": "./src/mcp/index.ts"
